requirements.txt
azure-ai-projects>=2.0.0
azure-identity~=1.25.2
ansible-core~=2.19.7
openai~=2.8.0
Flask~=3.1.0
python-dotenv~=1.1.0
